Support for a VUV photon visibility library from binary file.
There are some choices that are mostly workaround here.
This is technically a breaking change in that one of the data types
in the interface of PhotonVisibilityService has changed.
That type is aliased, and its interface is equivalent to the old one,
so no effect on code using PhotonVisibilityLibrary is expected.
IPhotonLibrary interface also changed, and more radically so.
There are not many use cases where this interface is addressed directly,
but for them some update might be needed.